Modify AccessComponent (2) tests to target Simplify, then merge them into SimplifyTest.cs.
